+++ local-cluster/src/validator_configs.rs
@@ -74,7 +74,6 @@ pub fn safe_clone_config(config: &ValidatorConfig) -> ValidatorConfig {
         unified_scheduler_handler_threads: config.unified_scheduler_handler_threads,
         ip_echo_server_threads: config.ip_echo_server_threads,
         rayon_global_threads: config.rayon_global_threads,
-        replay_forks_threads: config.replay_forks_threads,
         replay_transactions_threads: config.replay_transactions_threads,
         tvu_shred_sigverify_threads: config.tvu_shred_sigverify_threads,
         tvu_bls_sigverify_threads: config.tvu_bls_sigverify_threads,